fix(mysql): harden routine text protocol classification
