+++ src-tauri/src/drivers/mysql/tests.rs
@@ -609,16 +609,24 @@ fn routes_mysql_routine_ddl_through_text_protocol() {
         "DROP PROCEDURE IF EXISTS sociedades_close;",
         "CREATE PROCEDURE sociedades_close() SELECT 1;",
         "CREATE DEFINER=`root`@`localhost` PROCEDURE sociedades_close() SELECT 1;",
+        "CREATE  DEFINER=`root`@`localhost` PROCEDURE sociedades_close() SELECT 1;",
         "CREATE DEFINER=`root`@`localhost`   PROCEDURE sociedades_close() SELECT 1;",
         "CREATE OR REPLACE PROCEDURE sociedades_close() SELECT 1;",
+        "CREATE OR  REPLACE PROCEDURE sociedades_close() SELECT 1;",
+        "CREATE OR  REPLACE DEFINER=`root`@`localhost` PROCEDURE sociedades_close() SELECT 1;",
+        "CREATE OR REPLACE  DEFINER=`root`@`localhost` PROCEDURE sociedades_close() SELECT 1;",
         "CREATE OR REPLACE DEFINER=`root`@`localhost` PROCEDURE sociedades_close() SELECT 1;",
         "CREATE OR REPLACE DEFINER=`root`@`localhost`   PROCEDURE sociedades_close() SELECT 1;",
         "ALTER PROCEDURE sociedades_close COMMENT 'patched';",
         "DROP FUNCTION IF EXISTS sociedades_total;",
         "CREATE FUNCTION sociedades_total() RETURNS INT RETURN 1;",
         "CREATE DEFINER=`root`@`localhost` FUNCTION sociedades_total() RETURNS INT RETURN 1;",
+        "CREATE  DEFINER=`root`@`localhost` FUNCTION sociedades_total() RETURNS INT RETURN 1;",
         "CREATE DEFINER=`root`@`localhost`   FUNCTION sociedades_total() RETURNS INT RETURN 1;",
         "CREATE OR REPLACE FUNCTION sociedades_total() RETURNS INT RETURN 1;",
+        "CREATE OR  REPLACE FUNCTION sociedades_total() RETURNS INT RETURN 1;",
+        "CREATE OR  REPLACE DEFINER=`root`@`localhost` FUNCTION sociedades_total() RETURNS INT RETURN 1;",
+        "CREATE OR REPLACE  DEFINER=`root`@`localhost` FUNCTION sociedades_total() RETURNS INT RETURN 1;",
         "CREATE OR REPLACE DEFINER=`root`@`localhost` FUNCTION sociedades_total() RETURNS INT RETURN 1;",
         "CREATE OR REPLACE DEFINER=`root`@`localhost`   FUNCTION sociedades_total() RETURNS INT RETURN 1;",
         "ALTER FUNCTION sociedades_total COMMENT 'patched';",
